requirements.txt
streamlit==1.35.0
plotly==5.22.0
altair==5.3.0
streamlit-option-menu==0.3.13
numerize==0.12
Django==5.0.6
djangorestframework==3.15.1
keyboard==0.13.5
mysql-connector-python==8.4.0
psutil==6.0.0
PyMySQL==1.1.1
pandas==2.2.2
numpy==1.26.4